Restructure, adapted scoring
