+++ src/omop_graph/reasoning/resolvers/resolver_pipeline.py
@@ -1,8 +1,47 @@
 from dataclasses import dataclass
+from omop_graph.graph.kg import KnowledgeGraph
 
-from omop_graph.graph import KnowledgeGraph
-from .base import CandidateResolver, CandidateHit, ResolverConfidence
+from dataclasses import dataclass
+from typing import Optional, Iterable
+from .resolvers import CandidateResolver, ResolverConfidence, CandidateHit
+from ...graph.paths import GraphPath, find_shortest_paths, find_shortest_paths_dijkstra
+from ...graph.kg import KnowledgeGraph
+from ...graph.edges import PredicateKind
+from ...graph.scoring import PathProfile, rank_paths, path_profile
 
+import logging
+logger = logging.getLogger(__name__)
+
+@dataclass(frozen=True)
+class GroundingCandidate:
+    concept_id: int
+    is_standard: bool
+    label: str
+    reasons: tuple[str, ...]
+    confidence: ResolverConfidence
+
+    best_path_profile: Optional[PathProfile]
+    paths: Optional[tuple[GraphPath, ...]]
+
+    @property
+    def rank(self) -> tuple:
+        # Lower is better 
+
+        # Rank first based on confidence and whether the concept is a standard concept
+        rank = (self.confidence, not self.is_standard)
+
+        if self.best_path_profile is not None:
+            rank += self.best_path_profile.path_rank()
+
+        return rank
+
+@dataclass(frozen=True)
+class GroundingConstraints:
+    parent_ids: Optional[tuple[int, ...]]
+    allowed_domains: Optional[tuple[str, ...]]
+    allowed_vocabularies: Optional[tuple[str, ...]] = None
+    require_standard: bool = False
+    max_depth: int = 6
 
 @dataclass
 class ResolverPipeline:
@@ -14,7 +53,8 @@ class ResolverPipeline:
         *,
         stop_after_confidence: ResolverConfidence | None = None,
     ):
-        self.resolvers = resolvers
+        # Sort the resolvers by confidence so the stop logic works correctly
+        self.resolvers = tuple(sorted(resolvers, key=lambda r: r.confidence.value))
         self.stop_after_confidence = stop_after_confidence
 
     def resolve(
@@ -46,3 +86,114 @@ class ResolverPipeline:
                     results.append(hit)
 
         return results
+    
+    def ground_term(
+        self,
+        kg: KnowledgeGraph,
+        text: str,
+        *,
+        constraints: GroundingConstraints,
+    ) -> list[GroundingCandidate]:
+
+        results: list[GroundingCandidate] = []
+
+        for hit in self.resolve(kg, text):
+            ok, reasons = self._passes_constraints(kg, hit.concept_id, constraints)
+            if not ok:
+                continue
+            
+            if constraints.parent_ids is not None:
+                paths = self._find_hierarchy_paths(
+                    kg,
+                    hit.concept_id,
+                    constraints.parent_ids,
+                    max_depth=constraints.max_depth,
+                )
+                if not paths:
+                    continue  # fails hierarchy constraint
+                best_path_profile = self._best_profile(kg=kg, paths=paths)
+                paths = tuple(paths)
+            else:
+                # Placeholder with dummy values?
+                best_path_profile = None
+                paths = None
+
+            c = kg.concept_view(hit.concept_id)
+
+            results.append(
+                GroundingCandidate(
+                    concept_id=hit.concept_id,
+                    label=c.concept_name,
+                    best_path_profile=best_path_profile,
+                    reasons=tuple(reasons),
+                    paths=paths,
+                    confidence=hit.resolver_confidence,
+                    is_standard=c.standard_concept == "S",
+                )
+            )
+
+        results.sort(key=lambda r: r.rank)
+        return results
+ 
+    
+    @staticmethod
+    def _best_profile(
+        kg: KnowledgeGraph,
+        paths: list[GraphPath],
+    ) -> PathProfile:
+        profiles = [path_profile(kg, p) for p in paths]
+        return min(profiles)
+    
+    @staticmethod
+    def _find_hierarchy_paths(
+        kg: KnowledgeGraph,
+        concept_id: int,
+        parent_ids: tuple[int, ...],
+        *,
+        max_depth: int,
+        max_paths: int = 3,
+    ) -> list[GraphPath]:
+        paths = []
+
+        for parent in parent_ids:
+            found, trace = find_shortest_paths(
+                kg,
+                source=concept_id,
+                target=parent,
+                predicate_kinds={PredicateKind.ONTOLOGICAL},
+                max_depth=max_depth,
+                max_paths=max_paths,
+            )
+            paths.extend(found)
+
+        return paths
+    
+    @staticmethod
+    def _passes_constraints(
+        kg: KnowledgeGraph,
+        concept_id: int,
+        constraints: GroundingConstraints,
+    ) -> tuple[bool, list[str]]:
+        reasons = []
+
+        c = kg.concept_view(concept_id)
+
+        # domain constraint
+        if constraints.allowed_domains is not None:
+            if c.domain_id not in constraints.allowed_domains:
+                return False, [
+                    f"domain {c.domain_id} not in {constraints.allowed_domains}"
+                ]
+
+        # vocabulary constraint
+        if constraints.allowed_vocabularies:
+            if c.vocabulary_id not in constraints.allowed_vocabularies:
+                return False, [
+                    f"vocabulary {c.vocabulary_id} not allowed"
+                ]
+
+        # standardness
+        if constraints.require_standard and c.standard_concept is None:
+            return False, ["concept is non-standard"]
+
+        return True, reasons
